+++ apps/_root/app/swap2/layout.tsx
@@ -1,7 +1,9 @@
+import { Providers } from './providers'
+
 export const metadata = {
   title: 'SushiSwap 🍣 | Sushi',
 }
 
 export default function SwapLayout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>
+  return <Providers>{children}</Providers>
 }