+++ packages/react-query/src/index.ts
@@ -1,2 +1,5 @@
 export * from './hooks'
 export * from './PersistQueryClientProvider'
+
+// Re-export react-query
+// export * from '@tanstack/react-query'