+++ packages/ui/skeleton/index.ts
@@ -3,6 +3,9 @@ import { FC, HTMLProps } from 'react'
 import { Box } from './Box'
 import { Circle, CircleProps } from './Circle'
 
+/**
+ * @deprecated
+ */
 export const Skeleton: {
   Box: FC<HTMLProps<HTMLDivElement>>
   Circle: FC<CircleProps>