+++ packages/v3-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sushiswap/v3-sdk
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [813200e55]
+  - @sushiswap/currency@0.0.4
+  - @sushiswap/amm@0.1.4
+  - @sushiswap/tines@1.0.4
+
 ## 1.0.0
 
 ### Major Changes